Backend/requirements.txt
# Core LangChain library
langchain==0.3.26
# Web framework
flask==3.1.1 
# OpenAI API client
sentence-transformers==4.1.0
# PDF manipulation library
pypdf==5.6.1 
# Load environment variables from a .env file
python-dotenv==1.1.0
# Pinecone integration for LangChain
langchain-pinecone==0.2.8 
# OpenAI integration for LangChain
langchain-openai==0.3.24
# community contributions for langchain
langchain-community==0.3.26
# Install the package in editable mode
-e .